Use uint32s for delay, so you can delay longer than 65s

+++ PeriodicTask.cpp
@@ -2,17 +2,17 @@
 	PeriodicTask - ...
 	Copyright (C) 2016 Michael Adams (www.michael.net.nz)
 	All rights reserved.
-	
+
 	Permission is hereby granted, free of charge, to any person obtaining a
 	copy of this software and associated documentation files (the "Software"),
 	to deal in the Software without restriction, including without limitation
 	the rights to use, copy, modify, merge, publish, distribute, sublicense,
 	and/or sell copies of the Software, and to permit persons to whom the
 	Software is furnished to do so, subject to the following conditions:
-	
+
 	The above copyright notice and this permission notice shall be included
 	in all copies or substantial portions of the Software.
-	
+
 	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
 	OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 	FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -28,7 +28,7 @@
 #include "PeriodicTask.h"
 #include "Arduino.h"
 
-PeriodicTask::PeriodicTask(uint16_t period)
+PeriodicTask::PeriodicTask(uint32_t period)
 : period(period)
 , _next_tick(period)
 {
@@ -60,7 +60,7 @@ void PeriodicTask::next_run_at(uint32_t next_tick)
 }
 
 // when to run the task again (relative time)
-void PeriodicTask::next_run_in(uint16_t delay)
+void PeriodicTask::next_run_in(uint32_t delay)
 {
 	_next_tick = millis() + delay;
 }
